+++ ludwig/globals.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ==============================================================================
 
-LUDWIG_VERSION = "0.13.0"
+LUDWIG_VERSION = "0.14.1"
 
 MODEL_FILE_NAME = "model"
 MODEL_WEIGHTS_FILE_NAME = "model_weights"  # legacy pickle format
@@ -31,7 +31,7 @@ DESCRIPTION_FILE_NAME = "description.json"
 PREDICTIONS_PARQUET_FILE_NAME = "predictions.parquet"
 PREDICTIONS_SHAPES_FILE_NAME = "predictions.shapes.json"
 
-TRAINING_PREPROC_FILE_NAME = "training.hdf5"
+TRAINING_PREPROC_FILE_NAME = "training.parquet"
 
 HYPEROPT_STATISTICS_FILE_NAME = "hyperopt_statistics.json"